fixed bug for getDescendantsInFixedOrder, refactored method names

+++ appinventor/blocklyeditor/src/undoHandler.js
@@ -13,8 +13,8 @@
  *  - refactor variable and method names
  *  - implement for disconnections in case of healStack when DELETED_BY_KEY
  *  - fix when connecting into middle (if undo is done, then the bottom should connect back to top)
- *  - automatic block movements when superior connecting to inferior
- *  - redo
+ *  - recover from automatic block movements when superior block connects to inferior block
+ *  - implement redo
  *
  *  [Potential]
  *  - implement for block collapse/disable/add_comment changes (context menu orders)?
@@ -50,19 +50,17 @@ Blockly.UndoHandler.retrieveState = function () {
 };
 
 Blockly.UndoHandler.processState = function(state) {
-    var block = state.BLOCK;
-
     // STATE CHANGE TYPE: deleted -> revive what was deleted
     if(state.hasOwnProperty(Blockly.UndoHandler.STATE_TYPE_DELETED)) {
         var deletedXmlBlock = state[Blockly.UndoHandler.STATE_TYPE_DELETED].xmlBlock;
         Blockly.mainWorkspace.paste(deletedXmlBlock);
-        block = Blockly.selected;
+        state.BLOCK = Blockly.selected;
 
         // remap ids since they've changed if deleted and revived back
-        var deletedIds = state[Blockly.UndoHandler.STATE_TYPE_DELETED].deletedIds;
-        var revivedIds = Blockly.UndoHandler.getIdsForDescendants(block);
-        for(var i = 0; i < deletedIds.length; i++) {
-            Blockly.UndoHandler.remapDeletedBlockToRevivedBlock(deletedIds[i], revivedIds[i]);
+        var deletedBlocks = state[Blockly.UndoHandler.STATE_TYPE_DELETED].deletedBlocks;
+        var revivedBlocks = Blockly.UndoHandler.getDescendantsInFixedOrder(state.BLOCK);
+        for(var i = 0; i < deletedBlocks.length; i++) {
+            Blockly.UndoHandler.remapDeletedBlockToRevivedBlock(deletedBlocks[i], revivedBlocks[i]);
         }
     }
 
@@ -70,12 +68,23 @@ Blockly.UndoHandler.processState = function(state) {
     if(state.hasOwnProperty(Blockly.UndoHandler.STATE_TYPE_CONNECTED)) {
         var connectedConnectionIndices = state[Blockly.UndoHandler.STATE_TYPE_CONNECTED];
         for(var i = 0; i < connectedConnectionIndices.length; i++) {
-            var connection = block.getConnections_()[connectedConnectionIndices[i]];
-            if(connection.isSuperior()) {
-                connection.targetConnection.sourceBlock_.unplug();
+            var connection = state.BLOCK.getConnections_()[connectedConnectionIndices[i]];
+            if(connection.targetConnection != null) {
+                if (connection.isSuperior()) {
+                    connection.targetConnection.sourceBlock_.unplug();
+                }
+                else {
+                    connection.sourceBlock_.unplug();
+                }
             }
             else {
-                connection.sourceBlock_.unplug();
+                // might be null if failed to reconnect during previous state retrieval
+                // case: [1] and [2] were connected vertically, and [3] was connected in middle.
+                //       this makes [3] have connection records to [1] and [2].
+                //       however, if [2] is deleted and revived, [1] and [2] are already reconnected due to healing,
+                //       so undoing the disconnection while reviving [3] couldn't get finished.
+                //       since [3]'s connection to [1] and [2] were lost, it can't disconnect from them
+                console.log("Blockly.UndoHandler.processState: couldn't undo connection since connection was already lost");
             }
         }
     }
@@ -83,7 +92,7 @@ Blockly.UndoHandler.processState = function(state) {
     // STATE CHANGE TYPE: moved -> move back to last position
     if(state.hasOwnProperty(Blockly.UndoHandler.STATE_TYPE_MOVED)) {
         var lastPosition = state[Blockly.UndoHandler.STATE_TYPE_MOVED];
-        block.moveTo(lastPosition.x, lastPosition.y);
+        state.BLOCK.moveTo(lastPosition.x, lastPosition.y);
     }
 
     // STATE CHANGE TYPE: disconnected -> connect what was disconnected
@@ -96,22 +105,22 @@ Blockly.UndoHandler.processState = function(state) {
             var blocksConnectionToDisconnectedBlockIndex = disconnectedBlockRecord.blocksConnectionToDisconnectedBlockIndex;
             // only connect if disconnectedBlock didn't connect to something in the meanwhile (e.g. healStack) 
             if(disconnectedBlock.getConnections_()[disconnectedBlocksConnectionToBlockIndex].targetConnection == null) {
-                disconnectedBlock.getConnections_()[disconnectedBlocksConnectionToBlockIndex].connect(block.getConnections_()[blocksConnectionToDisconnectedBlockIndex]);
+                disconnectedBlock.getConnections_()[disconnectedBlocksConnectionToBlockIndex].connect(state.BLOCK.getConnections_()[blocksConnectionToDisconnectedBlockIndex]);
+            }
+            else {
+                // might be case where targetBlock was already connected to something else (such as connection healing) 
+                console.log("Blockly.UndoHandler.processState: couldn't undo disconnection since connection was already lost");
             }
         }
     }
 
     // STATE CHANGE TYPE: created -> dispose what was created
     if(state.hasOwnProperty(Blockly.UndoHandler.STATE_TYPE_CREATED)) {
-        block.dispose(false, false);
+        state.BLOCK.dispose(false, false);
     }
 
     // notify state change to undoIndicator
     Blockly.UndoHandler.notifyStateChange();
-};
-
-Blockly.UndoHandler.notifyStateChange = function () {
-    Blockly.mainWorkspace.undoIndicator.updateUndoIndicator(Blockly.UndoHandler.savedStates.length);
 };
 
 Blockly.UndoHandler.clearSavedStates = function () {
@@ -119,65 +128,21 @@ Blockly.UndoHandler.clearSavedStates = function () {
     Blockly.mainWorkspace.undoIndicator.updateUndoIndicator(Blockly.UndoHandler.savedStates.length);
 };
 
-Blockly.UndoHandler.remapDeletedBlockToRevivedBlock = function(deletedId, revivedId) {
-    for(var i = 0; i < Blockly.UndoHandler.savedStates.length; i++) {
-        var savedState = Blockly.UndoHandler.savedStates[i];
-        if(savedState.BLOCK && savedState.BLOCK.id == deletedId) {
-            savedState.BLOCK = Blockly.mainWorkspace.getBlockById(revivedId);
-        }
-        if(savedState[Blockly.UndoHandler.STATE_TYPE_DISCONNECTED]) {
-            var disconnectedBlocksRecords = savedState[Blockly.UndoHandler.STATE_TYPE_DISCONNECTED];
-            for (var j = 0; j < disconnectedBlocksRecords.length; j++) {
-                if (disconnectedBlocksRecords[j].disconnectedBlock.id == deletedId) {
-                    disconnectedBlocksRecords[j].disconnectedBlock = Blockly.mainWorkspace.getBlockById(revivedId);
-                }
-            }
-        }
-    }
-};
-
-// created own function, since order of getDescendants() differs depending on block connection order
-// this will always return fixed order of a block structure's descendants, regardless of in what order the blocks were connected 
-Blockly.UndoHandler.getIdsForDescendants = function(block) {
-    var idsForDescendants = [block.id];
-    var blockIds = [block.id];
-    
-    while(blockIds.length > 0) {
-        var nextBlockId = blockIds.pop();
-        var nextBlock = Blockly.mainWorkspace.getBlockById(nextBlockId);
-        var nextBlockConnections = nextBlock.getConnections_();
-        for(var i = 0; i < nextBlockConnections.length; i++) {
-            if(nextBlockConnections[i].targetConnection != null) {
-                var targetBlock = nextBlockConnections[i].targetConnection.sourceBlock_;
-                if(idsForDescendants.indexOf(targetBlock.id) < 0) {
-                    idsForDescendants.push(targetBlock.id);
-                    blockIds.push(targetBlock.id);
-                }
-            }
-        }
-    }
-    
-    return idsForDescendants;
-};
-
 Blockly.UndoHandler.createDeletedBlockRecord = function(deletionType) {
     var block = Blockly.UndoHandler.savedState.BLOCK;
     var xmlBlock = Blockly.Xml.blockToDom_(block);
-    var deletedIds = Blockly.UndoHandler.getIdsForDescendants(block);
+    var deletedBlocks = Blockly.UndoHandler.getDescendantsInFixedOrder(block);
 
-    // if it was dragged, all other horizontal children were dragged with it, so don't do deleteNext
-    // if it was deleted with the del key, horizontal children weren't deleted, so do deleteNext
-    if(deletionType == Blockly.UndoHandler.DELETED_BY_KEY) {
-        // if there were nextBlocks, those ids should be removed from deletedIds
-        if(block.getNextBlock()) {
-            var idsForNextBlockDescendants = Blockly.UndoHandler.getIdsForDescendants(block.getNextBlock());
-            for(var i = 0; i < idsForNextBlockDescendants.length; i ++) {
-                if(deletedIds.indexOf(idsForNextBlockDescendants[i]) > -1) {
-                    deletedIds.splice(deletedIds.indexOf(idsForNextBlockDescendants[i]), 1);
-                }
+    // if deleted by dragging, all other nextBlocks were dragged with it, so no special care needed
+    // if deleted by keyboard, nextBlocks aren't deleted, so remove nextBlocks from deletedBlocks
+    if(deletionType == Blockly.UndoHandler.DELETED_BY_KEY && block.getNextBlock()) {
+        var idsForNextBlockDescendants = Blockly.UndoHandler.getDescendantsInFixedOrder(block.getNextBlock());
+        for(var i = 0; i < idsForNextBlockDescendants.length; i ++) {
+            if(deletedBlocks.indexOf(idsForNextBlockDescendants[i]) > -1) {
+                deletedBlocks.splice(deletedBlocks.indexOf(idsForNextBlockDescendants[i]), 1);
             }
         }
-        // delete the next blocks from the XML representation too 
+        // delete nextBlocks from the XML representation that is gonna be saved too 
         Blockly.Xml.deleteNext(xmlBlock);
     }
 
@@ -186,7 +151,7 @@ Blockly.UndoHandler.createDeletedBlockRecord = function(deletionType) {
     xmlBlock.setAttribute('x', Blockly.RTL ? -xy.x : xy.x);
     xmlBlock.setAttribute('y', xy.y);
 
-    return {xmlBlock: xmlBlock, deletedIds: deletedIds};
+    return {xmlBlock: xmlBlock, deletedBlocks: deletedBlocks};
 };
 
 Blockly.UndoHandler.createMovedBlockRecord = function () {
@@ -194,14 +159,13 @@ Blockly.UndoHandler.createMovedBlockRecord = function () {
 };
 
 Blockly.UndoHandler.createConnectedBlockRecord = function(previousConnectedConnections) {
-    // check what new connected connections the block has made
-    var newConnectedConnectionsIndices = [];
+    var newConnectedConnectionsIndices = []; // check what new connected connections the block has made
     var block = Blockly.UndoHandler.savedState.BLOCK;
     
     var allConnections = block.getConnections_();
     for(var i = 0; i < allConnections.length; i++) {
       if(allConnections[i].targetConnection != null && previousConnectedConnections.indexOf(allConnections[i]) < 0) {
-          newConnectedConnectionsIndices.push(i); // save only index, since connection will become outdated when deletion happens
+          newConnectedConnectionsIndices.push(i); // save connection index only, since connection will be outdated when block is deleted
       }
     }
     
@@ -215,28 +179,18 @@ Blockly.UndoHandler.createDisconnectedBlockRecord = function(disconnectedBlocks)
         var disconnectedBlock = disconnectedBlocks[i];
         disconnectedBlocksRecords.push({
             disconnectedBlock: disconnectedBlock,
-            disconnectedBlocksConnectionToBlockIndex: Blockly.UndoHandler.getConnectionToTargetBlockIndex(disconnectedBlock, block),
-            blocksConnectionToDisconnectedBlockIndex: Blockly.UndoHandler.getConnectionToTargetBlockIndex(block, disconnectedBlock)
+            disconnectedBlocksConnectionToBlockIndex: Blockly.UndoHandler.getIndexOfConnectionToTargetBlock(disconnectedBlock, block),
+            blocksConnectionToDisconnectedBlockIndex: Blockly.UndoHandler.getIndexOfConnectionToTargetBlock(block, disconnectedBlock)
         });
     }
     
     return disconnectedBlocksRecords;
 };
 
-Blockly.UndoHandler.getConnectionToTargetBlockIndex = function(sourceBlock, targetBlock) {
-    var sourceBlockConnections = sourceBlock.getConnections_();
-    for(var i = 0; i < sourceBlockConnections.length; i++) {
-        if(sourceBlockConnections[i].targetConnection && (sourceBlockConnections[i].targetConnection.sourceBlock_ == targetBlock)) {
-            return i;
-        }
-    }
-};
-
 Blockly.UndoHandler.startRecord = function(block) {
     // not going to save states for changes in other workspaces (e.g. mutator workspaces, etc.)
     if(block.workspace == Blockly.mainWorkspace && Blockly.UndoHandler.isRecording == false) {
-        // reset savedState
-        Blockly.UndoHandler.savedState = {};
+        Blockly.UndoHandler.savedState = {}; // reset savedState
         Blockly.UndoHandler.savedState.BLOCK = block;
         Blockly.UndoHandler.isRecording = true;
     }
@@ -282,4 +236,58 @@ Blockly.UndoHandler.endRecord = function () {
         }
         Blockly.UndoHandler.isRecording = false;
     }
+};
+
+Blockly.UndoHandler.remapDeletedBlockToRevivedBlock = function(deletedBlock, revivedBlock) {
+    for(var i = 0; i < Blockly.UndoHandler.savedStates.length; i++) {
+        var savedState = Blockly.UndoHandler.savedStates[i];
+        if(savedState.BLOCK && savedState.BLOCK == deletedBlock) {
+            savedState.BLOCK = revivedBlock;
+        }
+        if(savedState[Blockly.UndoHandler.STATE_TYPE_DISCONNECTED]) {
+            var disconnectedBlocksRecords = savedState[Blockly.UndoHandler.STATE_TYPE_DISCONNECTED];
+            for(var j = 0; j < disconnectedBlocksRecords.length; j++) {
+                if (disconnectedBlocksRecords[j].disconnectedBlock == deletedBlock) {
+                    disconnectedBlocksRecords[j].disconnectedBlock = revivedBlock;
+                }
+            }
+        }
+    }
+};
+
+// created own function, since order of getDescendants() differs depending on block connection order
+// this will always return fixed order of a block structure's descendants, regardless of in what order the blocks were connected 
+Blockly.UndoHandler.getDescendantsInFixedOrder = function(block) {
+    var descendantsInFixedOrder = [block];
+    var blocksToTraverse = [block];
+    var parentBlock = block.getParent(); // not gonna traverse upwards
+
+    while(blocksToTraverse.length > 0) {
+        var nextBlock = blocksToTraverse.pop();
+        var nextBlockConnections = nextBlock.getConnections_();
+        for(var i = 0; i < nextBlockConnections.length; i++) {
+            if(nextBlockConnections[i].targetConnection != null) {
+                var targetBlock = nextBlockConnections[i].targetConnection.sourceBlock_;
+                if(descendantsInFixedOrder.indexOf(targetBlock) < 0 && targetBlock != parentBlock) {
+                    descendantsInFixedOrder.push(targetBlock);
+                    blocksToTraverse.push(targetBlock);
+                }
+            }
+        }
+    }
+
+    return descendantsInFixedOrder;
+};
+
+Blockly.UndoHandler.getIndexOfConnectionToTargetBlock = function(sourceBlock, targetBlock) {
+    var sourceBlockConnections = sourceBlock.getConnections_();
+    for(var i = 0; i < sourceBlockConnections.length; i++) {
+        if(sourceBlockConnections[i].targetConnection && (sourceBlockConnections[i].targetConnection.sourceBlock_ == targetBlock)) {
+            return i;
+        }
+    }
+};
+
+Blockly.UndoHandler.notifyStateChange = function () {
+    Blockly.mainWorkspace.undoIndicator.updateUndoIndicator(Blockly.UndoHandler.savedStates.length);
 };
